playground: moved IAppDialogSettings to types' file
